backend/requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database - PostgreSQL
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Database - Neo4j
neo4j==5.14.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
authlib==1.3.0
python-multipart==0.0.6

# AI / LLM Integration
openai==1.3.5
anthropic==0.7.1
httpx==0.25.1

# Import/Export
rdflib==7.0.0
pandas==2.1.3
